+++ springrest/src/main/java/com/mazdausa/dao/DataDao.java
@@ -8,9 +8,9 @@ public interface DataDao {
 
 	public boolean addEntity(Employee employee) throws Exception;
 
-	public Employee getEntityById(long id) throws Exception;
+	public Employee getEntityById(int id) throws Exception;
 
 	public List<Employee> getEntityList() throws Exception;
 
-	public boolean deleteEntity(long id) throws Exception;
+	public boolean deleteEntity(int id) throws Exception;
 }
