+++ springrest/src/main/java/com/mazdausa/services/DataServices.java
@@ -7,9 +7,9 @@ import com.mazdausa.model.Employee;
 public interface DataServices {
 	public boolean addEntity(Employee employee) throws Exception;
 
-	public Employee getEntityById(long id) throws Exception;
+	public Employee getEntityById(int id) throws Exception;
 
 	public List<Employee> getEntityList() throws Exception;
 
-	public boolean deleteEntity(long id) throws Exception;
+	public boolean deleteEntity(int id) throws Exception;
 }
